API/requirements.txt
flask
flask_cors
passlib
psycopg2
pyjwt
requests